Couple fixes, use the correct type for the ingame mask clientside, and make srue that there are enough deathmatch starts (another limit elsewhere that I missed)

+++ src/o_client.h
@@ -33,7 +33,7 @@
 
 ENetHost *localclient;
 int localid;
-uint8_t inGameMask;
+uint16_t inGameMask;
 
 void CL_Connect (char *srv_hn);
 void CL_Loop(void);
